to split navbar
