--- public/catalog.json
@@ -1,46 +1,0 @@
-[
-  {
-    "_id": "5d99f68e419d040e6ec0f722c",
-    "id": "men",
-    "name": "Мужские",
-    "parentId": "null",
-    "imgUrl": "img/catalog/men.png",
-    "description": "A category, that represents products for men",
-    "level": 0,
-    "date": "2019-10-06T14:13:34.304Z",
-    "__v": 0
-  },
-  {
-    "_id": "5d99f68e419d040eec0f7262c",
-    "id": "women",
-    "name": "Женские",
-    "parentId": "null",
-    "imgUrl": "img/catalog/women.png",
-    "description": "A category, that represents products for women",
-    "level": 0,
-    "date": "2019-10-06T14:13:34.304Z",
-    "__v": 0
-  },
-  {
-    "_id": "5d99f68e4179d040eec0f722c",
-    "id": "kids",
-    "name": "Детские",
-    "parentId": "null",
-    "imgUrl": "img/catalog/kids.png",
-    "description": "A category, that represents products for kids",
-    "level": 0,
-    "date": "2019-10-06T14:13:34.304Z",
-    "__v": 0
-  },
-  {
-    "_id": "5d99f68e419d040e8ec0f722c",
-    "id": "accessories",
-    "name": "Аксесуары",
-    "parentId": "null",
-    "imgUrl": "img/catalog/accessories.png",
-    "description": "A category, that represents products accessories",
-    "level": 0,
-    "date": "2019-10-06T14:13:34.304Z",
-    "__v": 0
-  }
-]
