+++ src/components/NavBar/MyOrders/MyOrders.module.scss
@@ -1,0 +1,44 @@
+.iconListItem {
+  position: relative;
+  align-items: center;
+  display: flex;
+}
+
+.iconCart {
+  position: relative;
+  bottom: 10px;
+  height: 36px;
+  width: 24px;
+  line-height: 1.1;
+}
+
+.productToCart {
+  position: relative;
+  display: block;
+  text-align: center;
+  width: 25px;
+  bottom: 25px;
+  line-height: 1;
+  font-size: 14px;
+  font-weight: 700;
+  color: #fafa64;
+}
+
+.infoCart {
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  margin-left: 15px;
+}
+
+.menuOrderTitle {
+  white-space: nowrap;
+  margin: 0 0 7px;
+  font-size: 17px;
+  line-height: 1;
+  font-weight: 700;
+}
+
+.menuOrderPrice {
+  font-size: 14px;
+}
